chore: move LimitOrderList into a better spot
wip
